Exorcism in Document Integretions Tests
